Update RewardSystem.cs
Chest of Sending is part of the Gothic theme pack, not the Veteran Reward System.
http://www.uoguide.com/Chest_of_Sending

Ethereal Horse was moved on real UO from a 3rd year reward to a 1st year reward. On ServUO it is correctly located under 1st year rewards, but whoever moved it forgot to remove it from 3rd year rewards also. It appears twice.