Bug Fixes - One more attempt at Mini Champ System bug where champ doesnt spawn and the system gets hung up - Removed redundant OnEnter/OnExit methods in Region/Mobile:
Here is the problem with the Region issue.  I can see why the redundant checks were added, so it would be called at the appropriate timing, ie OnExit called before hte new region was set to the mobile, and OnEnter begin called aferwards.  The problem is, in Region.cs, it is needed there to be called during the appropriate child/parent regions as well.  This will need to be looked into in the near future.