Update XmlQuestLeaders.cs
Calling AddLabel with a null value for the string will result in a gump interning the value as 0, meaning the first string added to the gump (if any) will be displayed in its place. In this case, guilds that are null are replaced with the gumps' title string. This fixes that issue.